Now ships chases
Note that for the eagle its not a good idea, since it gets more damage and doesnt do more damage, so it should actually keep a distance rather than chasing. This could be achieved by making the nodes of the route more intelligent, making possible to wait for other ships and so.